Refactor the user_reports data transform code to another file
